requirements.txt
# Mask iteration web app, bundled SAM3 source, and Validate_tools dependencies.

# Web app / API clients
aiohttp>=3.9
openai>=1.30

# Core image / data handling
numpy>=1.26,<2
Pillow>=10.0
opencv-python>=4.8
pycocotools>=2.0

# PyTorch runtime used by SAM3 image inference is intentionally not pinned here.
# Install torch and torchvision separately for your CUDA / MPS / CPU environment.

# SAM3 runtime dependencies from third_party/sam3/pyproject.toml
timm>=1.0.17
tqdm>=4.66
ftfy==6.1.1
regex>=2023.0.0
iopath>=0.1.10
typing_extensions>=4.8
huggingface_hub>=0.23

# SAM3 model/config utilities used by bundled modules
fvcore==0.1.5.post20221221
hydra-core>=1.1
omegaconf>=2.1
einops>=0.6
wget>=3.0
scipy>=1.10
scikit-image>=0.21
matplotlib>=3.7

# Optional SAM3 training/eval modules present in the bundle import these.
submitit>=1.5
torchmetrics>=1.3

# Additional bundled SAM3 modules imported by image/video/visualization helpers.
psutil>=5.9
pandas>=2.0
scikit-learn>=1.3
